# openapi_faim.json
{"openapi":"3.1.0","info":{"title":"FAIM Backend","description":"Fractal Antisymmetric Inheritance Memory (FAIM) API","version":"1.0.0"},"paths":{"/api/v1/chat":{"post":{"summary":"Chat","description":"Main chat endpoint for FAIM Lab.\n\nSteps:\n1. Retrieve FAIM context via naive_used_nodes (temporary P4 retriever)\n2. Stream tokens from LLM\n3. Publish tokens to SSE for live UI updates\n4. Publish used_nodes to UI after chat completes","operationId":"chat_api_v1_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatReq"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResp"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/graphs/{graph_id}/subgraph/{node_id}":{"get":{"tags":["Graph"],"summary":"Api Subgraph","operationId":"api_subgraph_api_v1_graphs__graph_id__subgraph__node_id__get","parameters":[{"name":"graph_id","in":"path","required":true,"schema":{"type":"string","title":"Graph Id"}},{"name":"node_id","in":"path","required":true,"schema":{"type":"string","title":"Node Id"}},{"name":"depth","in":"query","required":false,"schema":{"type":"integer","default":2,"title":"Depth"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GraphSubgraph"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/graphs/{graph_id}/recent-used-nodes":{"get":{"tags":["Graph"],"summary":"Recent Used Nodes","operationId":"recent_used_nodes_api_v1_graphs__graph_id__recent_used_nodes_get","parameters":[{"name":"graph_id","in":"path","required":true,"schema":{"type":"string","title":"Graph Id"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/graphs/{graph_id}/nodes":{"get":{"tags":["Graph"],"summary":"Api Scan Nodes","operationId":"api_scan_nodes_api_v1_graphs__graph_id__nodes_get","parameters":[{"name":"graph_id","in":"path","required":true,"schema":{"type":"string","title":"Graph Id"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":250,"title":"Limit"}},{"name":"order","in":"query","required":false,"schema":{"type":"string","default":"id","title":"Order"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NodeScanItem"},"title":"Response Api Scan Nodes Api V1 Graphs  Graph Id  Nodes Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/graphs/{graph_id}/node/{node_id}":{"get":{"tags":["Graph"],"summary":"Api Node Detail","operationId":"api_node_detail_api_v1_graphs__graph_id__node__node_id__get","parameters":[{"name":"graph_id","in":"path","required":true,"schema":{"type":"string","title":"Graph Id"}},{"name":"node_id","in":"path","required":true,"schema":{"type":"string","title":"Node Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NodeDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/graphs/{graph_id}/metrics":{"get":{"tags":["Graph"],"summary":"Api Metrics","operationId":"api_metrics_api_v1_graphs__graph_id__metrics_get","parameters":[{"name":"graph_id","in":"path","required":true,"schema":{"type":"string","title":"Graph Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GraphMetrics"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/graphs/{graph_id}/node/{node_id}/vector_stats":{"get":{"tags":["Graph"],"summary":"Api Vector Stats Route","operationId":"api_vector_stats_route_api_v1_graphs__graph_id__node__node_id__vector_stats_get","parameters":[{"name":"graph_id","in":"path","required":true,"schema":{"type":"string","title":"Graph Id"}},{"name":"node_id","in":"path","required":true,"schema":{"type":"string","title":"Node Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VectorStats"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/graphs/{graph_id}/node/{node_id}/neighbors":{"get":{"tags":["Graph"],"summary":"Api Neighbors","operationId":"api_neighbors_api_v1_graphs__graph_id__node__node_id__neighbors_get","parameters":[{"name":"graph_id","in":"path","required":true,"schema":{"type":"string","title":"Graph Id"}},{"name":"node_id","in":"path","required":true,"schema":{"type":"string","title":"Node Id"}},{"name":"k","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"K"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NeighborInfo"},"title":"Response Api Neighbors Api V1 Graphs  Graph Id  Node  Node Id  Neighbors Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/graphs/{graph_id}/node/{node_id}/lineage":{"get":{"tags":["Graph"],"summary":"Api Lineage","operationId":"api_lineage_api_v1_graphs__graph_id__node__node_id__lineage_get","parameters":[{"name":"graph_id","in":"path","required":true,"schema":{"type":"string","title":"Graph Id"}},{"name":"node_id","in":"path","required":true,"schema":{"type":"string","title":"Node Id"}},{"name":"depth","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Depth"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LineageModel"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/graphs/{graph_id}/degree_distribution":{"get":{"tags":["Graph"],"summary":"Api Degree Distribution","operationId":"api_degree_distribution_api_v1_graphs__graph_id__degree_distribution_get","parameters":[{"name":"graph_id","in":"path","required":true,"schema":{"type":"string","title":"Graph Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"integer"},"title":"Response Api Degree Distribution Api V1 Graphs  Graph Id  Degree Distribution Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/graphs/{graph_id}/cluster_map":{"get":{"tags":["Graph"],"summary":"Api Cluster Graph","operationId":"api_cluster_graph_api_v1_graphs__graph_id__cluster_map_get","parameters":[{"name":"graph_id","in":"path","required":true,"schema":{"type":"string","title":"Graph Id"}},{"name":"clusters","in":"query","required":false,"schema":{"type":"integer","default":16,"title":"Clusters"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/graphs/{graph_id}/add":{"post":{"tags":["Graph"],"summary":"Api Add Fragment Route","operationId":"api_add_fragment_route_api_v1_graphs__graph_id__add_post","parameters":[{"name":"graph_id","in":"path","required":true,"schema":{"type":"string","title":"Graph Id"}},{"name":"text","in":"query","required":true,"schema":{"type":"string","title":"Text"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/benchmarks/{graph_id}/snapshot":{"post":{"tags":["Benchmarks"],"summary":"Snapshot","operationId":"snapshot_api_v1_benchmarks__graph_id__snapshot_post","parameters":[{"name":"graph_id","in":"path","required":true,"schema":{"type":"string","title":"Graph Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BenchmarkPoint"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/benchmarks/":{"get":{"tags":["Benchmarks"],"summary":"Get Benchmarks","operationId":"get_benchmarks_api_v1_benchmarks__get","parameters":[{"name":"graph_id","in":"query","required":true,"schema":{"type":"string","title":"Graph Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/benchmarks/{graph_id}":{"get":{"tags":["Benchmarks"],"summary":"History","operationId":"history_api_v1_benchmarks__graph_id__get","parameters":[{"name":"graph_id","in":"path","required":true,"schema":{"type":"string","title":"Graph Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BenchmarkPoint"},"title":"Response History Api V1 Benchmarks  Graph Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/health":{"get":{"summary":"Health","operationId":"health_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/stream":{"get":{"summary":"Stream","description":"Main SSE stream for FAIM Lab.\nDelivers:\n  - chat_token\n  - chat_used\n  - fig_delta\n  - future evolution events","operationId":"stream_api_v1_stream_get","parameters":[{"name":"graph_id","in":"query","required":true,"schema":{"type":"string","title":"Graph Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"BenchmarkPoint":{"properties":{"timestamp":{"type":"string","title":"Timestamp"},"nodes":{"type":"integer","title":"Nodes"},"cr":{"type":"number","title":"Cr"},"redundancy":{"type":"number","title":"Redundancy"},"drift":{"type":"number","title":"Drift"},"latency":{"$ref":"#/components/schemas/LatencyPoint"}},"type":"object","required":["timestamp","nodes","cr","redundancy","drift","latency"],"title":"BenchmarkPoint"},"ChatReq":{"properties":{"graph_id":{"type":"string","title":"Graph Id"},"message":{"type":"string","title":"Message"}},"type":"object","required":["graph_id","message"],"title":"ChatReq"},"ChatResp":{"properties":{"reply":{"type":"string","title":"Reply"},"used_nodes":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Used Nodes"}},"type":"object","required":["reply","used_nodes"],"title":"ChatResp"},"GraphMetrics":{"properties":{"node_count":{"type":"integer","title":"Node Count","default":0},"edge_count":{"type":"integer","title":"Edge Count","default":0},"compression_ratio":{"type":"number","title":"Compression Ratio","default":0.0},"redundancy":{"type":"number","title":"Redundancy","default":0.0},"drift":{"type":"number","title":"Drift","default":0.0},"retrieve_p50_ms":{"type":"number","title":"Retrieve P50 Ms","default":0.0},"retrieve_p95_ms":{"type":"number","title":"Retrieve P95 Ms","default":0.0}},"type":"object","title":"GraphMetrics","description":"Metrics used by:\n  - /graph/{graph_id}/metrics\n  - /api/v1/benchmarks/{graph_id}*"},"GraphSubgraph":{"properties":{"nodes":{"items":{"$ref":"#/components/schemas/NodeBase"},"type":"array","title":"Nodes"},"links":{"items":{"$ref":"#/components/schemas/LinkModel"},"type":"array","title":"Links"}},"type":"object","title":"GraphSubgraph","description":"Structure returned to the FIG 3D graph:\n  { nodes: [...], links: [...] }"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LatencyPoint":{"properties":{"retrieve_p50_ms":{"type":"number","title":"Retrieve P50 Ms","default":0.0},"retrieve_p95_ms":{"type":"number","title":"Retrieve P95 Ms","default":0.0}},"type":"object","title":"LatencyPoint"},"LineageModel":{"properties":{"lineage":{"items":{"type":"string"},"type":"array","title":"Lineage"}},"type":"object","required":["lineage"],"title":"LineageModel"},"LinkModel":{"properties":{"source":{"type":"string","title":"Source"},"target":{"type":"string","title":"Target"}},"type":"object","required":["source","target"],"title":"LinkModel"},"NeighborInfo":{"properties":{"id":{"type":"string","title":"Id"},"distance":{"type":"number","title":"Distance"}},"type":"object","required":["id","distance"],"title":"NeighborInfo"},"NodeBase":{"properties":{"id":{"type":"string","title":"Id"},"label":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Label","default":""},"vector":{"anyOf":[{"items":{"type":"number"},"type":"array"},{"type":"null"}],"title":"Vector"},"payload":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Payload"}},"type":"object","required":["id"],"title":"NodeBase"},"NodeDetail":{"properties":{"id":{"type":"string","title":"Id"},"label":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Label","default":""},"vector":{"anyOf":[{"items":{"type":"number"},"type":"array"},{"type":"null"}],"title":"Vector"},"payload":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Payload"},"parents":{"anyOf":[{"items":{"$ref":"#/components/schemas/ParentInfo"},"type":"array"},{"type":"null"}],"title":"Parents"},"fractions":{"anyOf":[{"items":{"type":"number"},"type":"array"},{"type":"null"}],"title":"Fractions"},"novelty":{"anyOf":[{},{"type":"null"}],"title":"Novelty"},"inherited":{"anyOf":[{},{"type":"null"}],"title":"Inherited"},"children":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Children"},"degree":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Degree"},"redundancy_score":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Redundancy Score"},"evolution_flags":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Evolution Flags"},"parent_distances":{"anyOf":[{"additionalProperties":{"type":"number"},"type":"object"},{"type":"null"}],"title":"Parent Distances"},"vector_stats":{"anyOf":[{"additionalProperties":{"type":"number"},"type":"object"},{"type":"null"}],"title":"Vector Stats"}},"type":"object","required":["id"],"title":"NodeDetail"},"NodeScanItem":{"properties":{"id":{"type":"string","title":"Id"},"degree":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Degree"}},"type":"object","required":["id"],"title":"NodeScanItem"},"ParentInfo":{"properties":{"id":{"type":"string","title":"Id"},"fraction":{"type":"number","title":"Fraction","default":1.0}},"type":"object","required":["id"],"title":"ParentInfo"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VectorStats":{"properties":{"norm":{"type":"number","title":"Norm"},"mean":{"type":"number","title":"Mean"},"std":{"type":"number","title":"Std"},"min":{"type":"number","title":"Min"},"max":{"type":"number","title":"Max"}},"type":"object","required":["norm","mean","std","min","max"],"title":"VectorStats"}}}}